# improve: speed up native EXIF parsing