--- a/mapillary_tools/geotag/geotag_images_from_exif.py
+++ b/mapillary_tools/geotag/geotag_images_from_exif.py
@@ -1,4 +1,3 @@
-import io
 import logging
 import typing as T
 from multiprocessing import Pool
@@ -6,32 +5,11 @@ from pathlib import Path
 
 from tqdm import tqdm
 
-from .. import exceptions, exif_write, geo, types, utils
+from .. import exceptions, geo, types, utils
 from ..exif_read import ExifRead, ExifReadABC
 from .geotag_from_generic import GeotagImagesFromGeneric
 
 LOG = logging.getLogger(__name__)
-
-
-def verify_image_exif_write(
-    metadata: types.ImageMetadata,
-    image_bytes: T.Optional[bytes] = None,
-) -> None:
-    if image_bytes is None:
-        edit = exif_write.ExifEdit(metadata.filename)
-    else:
-        edit = exif_write.ExifEdit(image_bytes)
-
-    # The cast is to fix the type error in Python3.6:
-    # Argument 1 to "add_image_description" of "ExifEdit" has incompatible type "ImageDescription"; expected "Dict[str, Any]"
-    edit.add_image_description(
-        T.cast(T.Dict, types.desc_file_to_exif(types.as_desc(metadata)))
-    )
-
-    # Possible errors thrown here:
-    # - struct.error: 'H' format requires 0 <= number <= 65535
-    # - piexif.InvalidImageDataError
-    edit.dump_image_bytes()
 
 
 class GeotagImagesFromEXIF(GeotagImagesFromGeneric):
@@ -84,29 +62,15 @@ class GeotagImagesFromEXIF(GeotagImagesFromGeneric):
         image_path: Path, skip_lonlat_error: bool = False
     ) -> types.ImageMetadataOrError:
         try:
-            # load the image bytes into memory to avoid reading it multiple times
             with image_path.open("rb") as fp:
-                image_bytesio = io.BytesIO(fp.read())
-
-            image_bytesio.seek(0, io.SEEK_SET)
-            exif = ExifRead(image_bytesio)
-
+                exif = ExifRead(fp)
             image_metadata = GeotagImagesFromEXIF.build_image_metadata(
                 image_path, exif, skip_lonlat_error=skip_lonlat_error
-            )
-
-            image_bytesio.seek(0, io.SEEK_SET)
-            verify_image_exif_write(
-                image_metadata,
-                image_bytes=image_bytesio.read(),
             )
         except Exception as ex:
             return types.describe_error_metadata(
                 ex, image_path, filetype=types.FileType.IMAGE
             )
-
-        image_bytesio.seek(0, io.SEEK_SET)
-        image_metadata.update_md5sum(image_bytesio)
 
         return image_metadata