--- a/tests/integration/test_history.py
+++ b/tests/integration/test_history.py
@@ -33,38 +33,6 @@ def test_upload_images(
         upload.remove()
 
     x = subprocess.run(
-        f"{EXECUTABLE} process --file_types=image {str(setup_data)}",
-        shell=True,
-    )
-    assert x.returncode == 0, x.stderr
-    verify_descs(
-        [
-            {
-                "error": {
-                    "message": "The image was already uploaded",
-                    "type": "MapillaryUploadedAlreadyError",
-                },
-                "filename": str(Path(setup_data.join("images").join("DSC00001.JPG"))),
-            },
-            {
-                "error": {
-                    "message": "The image was already uploaded",
-                    "type": "MapillaryUploadedAlreadyError",
-                },
-                "filename": str(Path(setup_data.join("images").join("DSC00497.JPG"))),
-            },
-            {
-                "error": {
-                    "message": "The image was already uploaded",
-                    "type": "MapillaryUploadedAlreadyError",
-                },
-                "filename": str(Path(setup_data.join("images").join("V0370574.JPG"))),
-            },
-        ],
-        Path(setup_data, "mapillary_image_description.json"),
-    )
-
-    x = subprocess.run(
         f"{EXECUTABLE} process_and_upload --file_types=image {UPLOAD_FLAGS} {str(setup_data)}",
         shell=True,
     )