- **disable metadata validation and upload status in process stage**
- **turn off detials and debug flags in exifread.process_file**
- **improve exif extraction**

Before: processing 9852 images
```
uv run -m mapillary_tools.commands process  --desc_path /tmp/foo.json   66.19s user 8.62s system 515% cpu 14.508 total
```

After: processing 9852 images (without validation)
```
uv run -m mapillary_tools.commands process  --desc_path /tmp/foo.json   5.98s user 2.37s system 238% cpu 3.494 total
```

After: processing 9852 images (with validation)
```
uv run -m mapillary_tools.commands process  --desc_path /tmp/foo.json   14.24s user 3.97s system 277% cpu 6.570 total
```